fix : 탐험 페이지 분할
- ExploreMain : 탐험 메인 페이지
- ExploreCamera : 촬영한 사진이 저장된 위치 가져오기
- ExploreDetail : 사진 위치 Text 로 나타냄
